+++ packages/desktop/src/global.d.ts
@@ -2,6 +2,7 @@ export declare global {
   interface Window {
     archives: {
       readSitePackagesSnapshot: () => Promise<Uint8Array>;
+      readRequirements: () => Promise<string[]>;
       readStreamlitAppDirectory: () => Promise<Record<string, Buffer>>;
     };
   }