+++ packages/mountable/src/index.tsx
@@ -49,6 +49,7 @@ export function mount(
   );
 
   const kernelWithToast = new StliteKernelWithToast(kernel);
+  let currentKernel: StliteKernel | StliteKernelWithToast = kernelWithToast;
 
   return {
     unmount: () => {
@@ -56,20 +57,26 @@ export function mount(
       ReactDOM.unmountComponentAtNode(container);
     },
     install: (requirements: string[]) => {
-      return kernelWithToast.install(requirements);
+      return currentKernel.install(requirements);
     },
     writeFile: (
       path: string,
       data: string | ArrayBufferView,
       opts?: Record<string, any>
     ) => {
-      return kernelWithToast.writeFile(path, data, opts);
+      return currentKernel.writeFile(path, data, opts);
     },
     renameFile: (oldPath: string, newPath: string) => {
-      return kernelWithToast.renameFile(oldPath, newPath);
+      return currentKernel.renameFile(oldPath, newPath);
     },
     unlink: (path: string) => {
-      return kernelWithToast.unlink(path);
+      return currentKernel.unlink(path);
+    },
+    enableToast: (): void => {
+      currentKernel = kernelWithToast;
+    },
+    disableToast: (): void => {
+      currentKernel = kernel;
     },
   };
 }
@@ -168,6 +175,58 @@ if (window.top) {
   window.top.postMessage(
     {
       streamlitstatus: "ready",
+      entrypoint: "streamlit_app.py",
+      files: {
+        "streamlit_app.py": `import streamlit as st
+import matplotlib.pyplot as plt
+import numpy as np
+import os
+
+st.markdown("Files in \`.\`:")
+st.write(os.listdir("."))
+
+st.markdown("\`./foo/foo.txt\`:")
+with open("./foo/foo.txt") as f:
+    st.write(f.read())
+
+st.markdown("\`./bar.txt\`:")
+with open("./bar.txt") as f:
+    st.write(f.read())
+
+size = st.slider("Sample size", 100, 1000)
+
+arr = np.random.normal(1, 1, size=size)
+fig, ax = plt.subplots()
+ax.hist(arr, bins=20)
+
+st.pyplot(fig)
+
+st.latex(r'''
+     a + ar + a r^2 + a r^3 + \\cdots + a r^{n-1} =
+     \\sum_{k=0}^{n-1} ar^k =
+     a \\left(\\frac{1-r^{n}}{1-r}\\right)
+     ''')
+
+import pandas as pd
+import numpy as np
+
+df = pd.DataFrame(
+     np.random.randn(1000, 2) / [50, 50] + [37.76, -122.4],
+     columns=['lat', 'lon'])
+
+st.map(df)
+    `,
+        "bar.txt": {
+          url: "/test-files/bar.txt",
+        },
+      },
+      archives: [
+        {
+          url: "/test-files/foo.zip",
+          format: "zip",
+        },
+      ],
+      requirements: ["matplotlib"],
     },
     "*"
   );