+++ packages/sharing-editor/public/samples/011_component_gallery/pages/data.dataframe1.py
@@ -1,11 +1,15 @@
-import streamlit as st
-import pandas as pd
 import numpy as np
+import pandas as pd
+import streamlit as st
+
 
 @st.cache_data
 def load_data():
-    df = pd.DataFrame(np.random.randn(10, 20), columns=("col %d" % i for i in range(20)))
+    df = pd.DataFrame(
+        np.random.randn(10, 20), columns=("col %d" % i for i in range(20))
+    )
     return df
+
 
 df = load_data()