+++ packages/sharing-editor/public/samples/011_component_gallery/pages/data.progress_column.py
@@ -1,0 +1,28 @@
+import pandas as pd
+import streamlit as st
+
+
+@st.cache_data
+def load_data():
+    return pd.DataFrame(
+        {
+            "sales": [200, 550, 1000, 80],
+        }
+    )
+
+
+data_df = load_data()
+
+st.data_editor(
+    data_df,
+    column_config={
+        "sales": st.column_config.ProgressColumn(
+            "Sales volume",
+            help="The sales volume in USD",
+            format="$%f",
+            min_value=0,
+            max_value=1000,
+        ),
+    },
+    hide_index=True,
+)